requirements.txt
asn1crypto==0.24.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
cmudict==0.4.1
cryptography==2.4.2
decorator==4.3.0
http-ece==1.0.5
idna==2.8
Mastodon.py==1.3.1
pronouncing==0.2.0
pycparser==2.19
python-dateutil==2.7.5
pytz==2018.9
requests==2.21.0
six==1.12.0
urllib3==1.24.1
